Rename terms for submission compliance from data

Change the variable names in the .df (only) and return that object to the user. This function simply calls xportr_tidy_names on colnames(.df)

xportr_df_varnames(
  .df,
  verbose = getOption("xportr.type_verbose", "none"),
  relo_2_end = TRUE,
  letter_for_num_prefix = "x",
  sep = "",
  replace_vec = c(`'` = "", `"` = "", `%` = "_pct_", `#` = "_nmbr_"),
  dict_dat = data.frame(original_varname = character(), dict_varname = character())
)

Arguments

.df
An R object with columns that can be coerced

verbose
The action the function takes when a variable isn't typed properly. Options are 'stop', 'warn', 'message', and 'none'

relo_2_end
logical, if TRUE: numerical prefix bundles will be relocated to the end of the string. A prefix bundle is determined as a number or "grouping of numbers separated by special characters and other punctuation". See details section for more info on prefix bundles.

letter_for_num_prefix
character. Will be ignored if relo_2_end = TRUE. Per CDISC & regulatory body requirements, variable names cannot start with a number, so if you want to leave the number of the left-hand side, use this argument to insert a starting letter before the numeric prefix.

sep
string of only one character long, intended to separate words from one another. In general, only "" and "" are possible, but the user can insert a letter if desired. Note that "" is discouraged because it is only permissible for legacy studies started on or before Dec 17, 2016.

replace_vec
A named character vector where the name is replaced by the value.

dict_dat
a data frame containing two variables: the original_varname and the dict_varname to find and replace

Value

Returns the modified table.

Details

Abbreviating variable names in the xportr pkg uses a step-by-step process detailed below. Each original variable name will be renamed based on the method performing the least amount of work to achieve the min character length requirement (8 chars), thus maintaining maximum originality. Therefore, the function only moves on to the next method on an as-needed basis; namely, when the term is still greater than 8 characters, even after lite modification.

(1) Blanks: Any columns that are missing a variable name (i.e., the header was blank in the source file) will be renamed to 'V' + the column position. So if the 2nd column is blank, it will receive a 'V2' rename.

(2) Use dictionary of controlled terminology: For example: 'Subject ID' may better be suited as 'SUBJID' within your organization. Note, that dictionary terms are expected to be submission compliant and will not be further abbreviated. They will, however, undergo a check for non-compliance.

(3) Do nothing! Or at the very least, mimic what SAS does automatically when cleaning up variable names during a PROC IMPORT. Namely, replace any special characters with underscores (''), capitalize everything, and if the value starts with a digit, add the '' prefix. If the 'SASified' name is <= 8 chars, then the function will use that rename. However, if its still too long, the function will try removing any extra special characters or spaces to help reduce to 8 chars.

(4) Find the STEM or ROOT word of each original variable name. For example, if the original contains the word 'resting', the 'ing' will be dropped and only the root word 'rest' will be considered. If less than 8 chars, the algorithm suggests that result. If its still too long, the function will, again, remove any special characters or spaces from the stemmed word(s).

(5) Apply an abbreviation algorithm who's primary goal is readability, such that the results remain unique. The methods described below are a bit more 'involved', but the results are very robust. First, you should know that characters are always stripped from the end of the strings first (i.e. from right to left). If an element of the variable name contains more than one word (words are separated by spaces) then at least one letter from each word will be retained.

Method: First spaces at the ends of the string are stripped. Then (if necessary) any other spaces are stripped. Next, lower case vowels are removed followed by lower case consonants. Finally if the abbreviation is still longer than 8 chars, upper case letters and symbols are stripped.

When identifying 'words', the app performs some pre-processing steps in the following order:

* Certain symbols are replaced. Like the '%' symbol is replaced with 'PCT'. Ex: 'I_Ate_%' becomes 'I_Ate_PCT'
* Replace any symbols with a blank space. Ex: 'I_Ate' becomes 'I Ate'
* Find when there are two capital letters next to each other, followed by a lower case letter then adds a space between the two capital letters to separate the assumed 'words'. Ex: 'IAte' becomes 'I Ate'
* Insert a space between a number followed by a character. Ex: 'iAte1meal' becomes 'iAte1 meal'
* Insert a space between a character followed by a number. Ex: 'iAte1meal' becomes 'iAte 1meal'

What do we abbreviate when? If a stemmed word exists, the app will apply the abbreviation algorithm described above on the stemmed version of the variable name, else the original variable name.

Since submission guidelines indicate variables may not start with a number, when found, the algorithm will either add and maintain a '_' "prefix bundle" through any transformations detailed above, or it will relocate the "prefix bundle" to the end of the term (the default behavior). What if a term starts with non-standard numerical prefix? Currently, the function accounts for the following types of prefix bundles, where 3 is used as an example starting digit:

* x = "3a. hey" will return "3a"
* x = "3_17a_hey" will return "3_17a"
* x = "3_17_hey" will return "3_17"
* x = "3_17hey" will return "3_17"
* x = "3hey" will return "3"

See also

Examples

vars <- c("", "STUDYID", "studyid", "subject id", "1c. ENT", "1b. Eyes",
     "1d. Lungs", "1e. Heart", "year number", "1a. Skin_Desc")
adxx <- data.frame(matrix(0, ncol = 10, nrow = 3))
colnames(adxx) <- vars

my_dictionary <- data.frame(original_varname = "subject id", dict_varname = "subjid")

metacore <- data.frame(
  dataset = "adxx",      # fix both here and in metacore?
  variable = c(vars, "NotUsed")
) %>%
dplyr::union(
   data.frame( dataset = "advs", variable = c("carrot", "cake"))
)

xportr_df_varnames(adxx) # default

#> 
#> The following variable name validation checks failed:
#> Must be 8 characters or less: Variables `subject id`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> Must start with a letter: Variables `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, and `1a. Skin_Desc`.
#> Cannot contain any non-ASCII, symbol or underscore characters: Variables `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> Cannot contain any lowercase characters Variables ``, `studyid`, `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.

#> 

#> 

#> -- 9 of 10 (90%) variables were renamed --

#> 

#> Var 1:  '' was renamed to 'V1'
#> Var 3:  'studyid' was renamed to 'STUDYID2'
#> Var 4:  'subject id' was renamed to 'SUBJECTD'
#> Var 5:  '1c. ENT' was renamed to 'ENT1C'
#> Var 6:  '1b. Eyes' was renamed to 'EYES1B'
#> Var 7:  '1d. Lungs' was renamed to 'LUNGS1D'
#> Var 8:  '1e. Heart' was renamed to 'HEART1E'
#> Var 9:  'year number' was renamed to 'YEARNUMB'
#> Var 10:  '1a. Skin_Desc' was renamed to 'SKNDSC1A'

#> 
#> All renamed variables passed validation.

#>   V1 STUDYID STUDYID2 SUBJECTD ENT1C EYES1B LUNGS1D HEART1E YEARNUMB SKNDSC1A
#> 1  0       0        0        0     0      0       0       0        0        0
#> 2  0       0        0        0     0      0       0       0        0        0
#> 3  0       0        0        0     0      0       0       0        0        0

xportr_df_varnames(adxx, relo_2_end = FALSE) # prefix numbers on left-hand side

#> 
#> The following variable name validation checks failed:
#> Must be 8 characters or less: Variables `subject id`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> Must start with a letter: Variables `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, and `1a. Skin_Desc`.
#> Cannot contain any non-ASCII, symbol or underscore characters: Variables `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> Cannot contain any lowercase characters Variables ``, `studyid`, `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.

#> 

#> -- 9 of 10 (90%) variables were renamed --

#> 

#> Var 1:  '' was renamed to 'V1'
#> Var 3:  'studyid' was renamed to 'STUDYID2'
#> Var 4:  'subject id' was renamed to 'SUBJECTD'
#> Var 5:  '1c. ENT' was renamed to 'X1CENT'
#> Var 6:  '1b. Eyes' was renamed to 'X1BEYES'
#> Var 7:  '1d. Lungs' was renamed to 'X1DLUNGS'
#> Var 8:  '1e. Heart' was renamed to 'X1EHEART'
#> Var 9:  'year number' was renamed to 'YEARNUMB'
#> Var 10:  '1a. Skin_Desc' was renamed to 'X1SKNDSC'

#> 
#> All renamed variables passed validation.

#>   V1 STUDYID STUDYID2 SUBJECTD X1CENT X1BEYES X1DLUNGS X1EHEART YEARNUMB
#> 1  0       0        0        0      0       0        0        0        0
#> 2  0       0        0        0      0       0        0        0        0
#> 3  0       0        0        0      0       0        0        0        0
#>   X1SKNDSC
#> 1        0
#> 2        0
#> 3        0

xportr_df_varnames(adxx, dict_dat  = my_dictionary) # 'SUBJID' used

#> 
#> The following variable name validation checks failed:
#> Must be 8 characters or less: Variables `subject id`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> Must start with a letter: Variables `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, and `1a. Skin_Desc`.
#> Cannot contain any non-ASCII, symbol or underscore characters: Variables `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> Cannot contain any lowercase characters Variables ``, `studyid`, `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.

#> 

#> -- 9 of 10 (90%) variables were renamed --

#> 

#> Var 1:  '' was renamed to 'V1'
#> Var 3:  'studyid' was renamed to 'STUDYID2'
#> Var 4:  'subject id' was renamed to 'SUBJID'
#> Var 5:  '1c. ENT' was renamed to 'ENT1C'
#> Var 6:  '1b. Eyes' was renamed to 'EYES1B'
#> Var 7:  '1d. Lungs' was renamed to 'LUNGS1D'
#> Var 8:  '1e. Heart' was renamed to 'HEART1E'
#> Var 9:  'year number' was renamed to 'YEARNUMB'
#> Var 10:  '1a. Skin_Desc' was renamed to 'SKNDSC1A'

#> 
#> All renamed variables passed validation.

#>   V1 STUDYID STUDYID2 SUBJID ENT1C EYES1B LUNGS1D HEART1E YEARNUMB SKNDSC1A
#> 1  0       0        0      0     0      0       0       0        0        0
#> 2  0       0        0      0     0      0       0       0        0        0
#> 3  0       0        0      0     0      0       0       0        0        0

xportr_df_varnames(adxx, sep = "_") # permissible for legacy studies

#> 
#> The following variable name validation checks failed:
#> Must be 8 characters or less: Variables `subject id`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> Must start with a letter: Variables `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, and `1a. Skin_Desc`.
#> Cannot contain any non-ASCII, symbol or underscore characters: Variables `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> Cannot contain any lowercase characters Variables ``, `studyid`, `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.

#> 

#> -- 9 of 10 (90%) variables were renamed --

#> 

#> Var 1:  '' was renamed to 'V1'
#> Var 3:  'studyid' was renamed to 'STUDYID2'
#> Var 4:  'subject id' was renamed to 'SUBJCTID'
#> Var 5:  '1c. ENT' was renamed to 'ENT_1C'
#> Var 6:  '1b. Eyes' was renamed to 'EYES_1B'
#> Var 7:  '1d. Lungs' was renamed to 'LUNGS_1D'
#> Var 8:  '1e. Heart' was renamed to 'HEART_1E'
#> Var 9:  'year number' was renamed to 'YEARNMBR'
#> Var 10:  '1a. Skin_Desc' was renamed to 'SKNDSC1A'

#> 
#> The following variable name validation checks still failed:
#> Cannot contain any non-ASCII, symbol or underscore characters: Variables `ENT_1C`, `EYES_1B`, `LUNGS_1D`, and `HEART_1E`.
#> Cannot contain any lowercase characters Variables `ENT_1C`, `EYES_1B`, `LUNGS_1D`, and `HEART_1E`.

#>   V1 STUDYID STUDYID2 SUBJCTID ENT_1C EYES_1B LUNGS_1D HEART_1E YEARNMBR
#> 1  0       0        0        0      0       0        0        0        0
#> 2  0       0        0        0      0       0        0        0        0
#> 3  0       0        0        0      0       0        0        0        0
#>   SKNDSC1A
#> 1        0
#> 2        0
#> 3        0