SAS format

Assigns SAS format from a variable level metadata to a given data frame.

xportr_format(
  .df,
  metacore,
  domain = NULL,
  verbose = getOption("xportr.format_verbose", "none")
)

Arguments

.df
A data frame of CDISC standard.

metacore
A data frame containing variable level metadata.

domain
A character value to subset the .df. If NULL(default), uses .df value as a subset condition.

verbose
The action the function takes when a variable label isn't. found. Options are 'stop', 'warn', 'message', and 'none'

Value

Data frame with SASformat attributes for each variable.

See also

Examples

adsl <- data.frame(
  USUBJID = c(1001, 1002, 1003),
  BRTHDT = c(1, 1, 2)
)

metacore <- data.frame(
  dataset = c("adsl", "adsl"),
  variable = c("USUBJID", "BRTHDT"),
  format = c(NA, "DATE9.")
)

adsl <- xportr_format(adsl, metacore)