Coerce variable type

Current assumptions: columns_meta is a data.frame with names "Variables", "Type"

xportr_type(
  .df,
  datadef,
  domain = NULL,
  verbose = getOption("xportr.type_verbose", "none")
)

Arguments

.df
An R object with columns that can be coerced

datadef
Either a data.frame that has the names of all possible columns and their types, or a DataDef object from the DataDef package. Required column names are dataset, variables, type

domain
Name of the dataset. Ex ADAE/DM. This will be used to subset the datadef object. If none is passed it is assumed to be the name of the dataset passed in .df.

verbose
The action the function takes when a variable isn't typed properly. Options are 'stop', 'warn', 'message', and 'none'

Value

Returns the modified table.

Examples

datadef <- data.frame(
  dataset = "test",
  variable = c("Subj", "Param", "Val", "NotUsed"),
  type = c("numeric", "character", "numeric", "character")
)

.df <- data.frame(
 Subj = as.character(123, 456, 789),
 Different = c("a", "b", "c"),
 Val = c("1", "2", "3"),
 Param = c("param1", "param2", "param3")
)

df2 <- xportr_type(.df, datadef, "test")

#> 

#> ── Variable type mismatches found. ──

#> 

#> ✔ 2 variables coerced