Write xpt v5 transport file

Writes a local data frame into SAS transport file of version 5. The SAS transport format is an open format, as is required for submission of the data to the FDA.

Usage

xportr_write(
  .df,
  path,
  metadata = NULL,
  domain = NULL,
  strict_checks = FALSE,
  label = deprecated()
)

Arguments

.df

A data frame to write.

path

Path where transport file will be written. File name sans will be used as xpt name.

metadata

A data frame containing variable level metadata. See 'Metadata' section for details.

domain

Appropriate CDSIC dataset name, e.g. ADAE, DM. Used to subset the metadata object. If none is passed, then xportr_metadata() must be called before hand to set the domain as an attribute of .df.

strict_checks

If TRUE, xpt validation will report errors and not write out the dataset. If FALSE, xpt validation will report warnings and continue with writing out the dataset. Defaults to FALSE

label

Previously used to to set the Dataset label. Use the metadata argument to set the dataset label.

Value

A data frame. xportr_write() returns the input data invisibly.

Details

Variable and dataset labels are stored in the "label" attribute.

SAS length are stored in the "SASlength" attribute.

SAS format are stored in the "SASformat" attribute.

SAS type are stored in the "SAStype" attribute.

Examples

adsl <- data.frame(
  SUBL = as.character(123, 456, 789),
  DIFF = c("a", "b", "c"),
  VAL = c("1", "2", "3"),
  PARAM = c("param1", "param2", "param3")
)

var_spec <- data.frame(
  dataset = "adsl",
  label = "Subject-Level Analysis Dataset",
  data_label = "ADSL"
)
xportr_write(adsl,
  path = paste0(tempdir(), "/adsl.xpt"),
  domain = "adsl",
  metadata = var_spec,
  strict_checks = FALSE
)